Create ReflectionPkg
The TestHarnEss class is used to instantiate classes, get names of all the methods, class and public and private members of TestClass.
The moto of the project is to use combinations of methods , up to four in this class. It is done by passing arguments to the methods and getting return values which is again input to another method and so on. Such combinations of inputs checks primitive types and return values.